coffee-store-backend/requirements.txt
# Web
fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
gunicorn>=21.0.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Auth & validation
python-jose[cryptography]>=3.3.0
passlib[argon2-cffi]>=1.7.4
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# HTTP client
httpx>=0.26.0


# Dev & test
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
respx>=0.20.0
black>=24.0.0
ruff>=0.2.0
mypy>=1.8.0
argon2_cffi
